# api/openapi.json
{
  "components": {
    "responses": {
      "BadRequest": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "The request was invalid."
      },
      "NotFound": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "The requested Spaniel resource was not found."
      }
    },
    "schemas": {
...
      "DashboardPanel": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "additionalProperties": false,
        "properties": {
          "dashboard_id": {
            "type": "string"
          },
          "display_type": {
            "enum": [
              "single_value",
              "time_series",
              "table",
              "heatmap",
              "entity_list",
              "trace_list",
              "span_list",
              "log_list",
              "deploy_correlation"
            ],
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "layout_json": {
            "type": "string"
          },
          "position": {
            "type": "number"
          },
          "query_sql": {
            "type": "string"
          },
          "query_version": {
            "type": "number"
          },
          "settings_json": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "updated_at": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "dashboard_id",
          "title",
          "display_type",
          "query_sql",
          "query_version",
          "settings_json",
          "layout_json",
          "position",
          "updated_at"
        ],
        "type": "object"
      },
...
      "Error": {
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "Ok": {
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "The complete local Spaniel HTTP API. JSON endpoints return the Envelope shape unless an operation explicitly returns an export.",
    "title": "Spaniel API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/dashboards": {
      "get": {
        "operationId": "listDashboards",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "example": {
                    "summary": "Representative Spaniel response",
                    "value": {
                      "data": [
                        {
                          "created_at": 1.5,
                          "description": "example",
                          "id": "session_2026-10-07_1200",
                          "name": "POST /checkout",
                          "panels": [
                            "…"
                          ],
                          "updated_at": 1.5,
                          "variables": [
                            "…"
                          ]
                        }
                      ]
                    }
                  }
                },
                "schema": {
                  "properties": {
                    "data": {
                      "items": {
                        "$ref": "#/components/schemas/Dashboard"
                      },
                      "type": "array"
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "summary": "list Dashboards",
        "tags": [
          "Dashboards"
        ]
      },
      "post": {
        "operationId": "createDashboard",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object"
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "example": {
                    "summary": "Representative Spaniel response",
                    "value": {
                      "data": {
                        "created_at": 1.5,
                        "description": "example",
                        "id": "session_2026-10-07_1200",
                        "name": "POST /checkout",
                        "panels": [
                          {
                            "dashboard_id": "…",
                            "display_type": "…",
                            "id": "…",
                            "layout_json": "…",
                            "query_sql": "…",
                            "query_version": "…",
                            "settings_json": "…",
                            "title": "…"
                          }
                        ],
                        "updated_at": 1.5,
                        "variables": [
                          {
                            "dashboard_id": "…",
                            "default_value": "…",
                            "kind": "…",
                            "name": "…",
                            "options_json": "…",
                            "source": "…"
                          }
                        ]
                      }
                    }
                  }
                },
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Dashboard"
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "summary": "create Dashboard",
        "tags": [
          "Dashboards"
        ]
      }
    },
    "/api/dashboards/import": {
      "post": {
        "operationId": "importDashboardConfig",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object"
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "example": {
                    "summary": "Representative Spaniel response",
                    "value": {
                      "data": {
                        "created_at": 1.5,
                        "description": "example",
                        "id": "session_2026-10-07_1200",
                        "name": "POST /checkout",
                        "panels": [
                          {
                            "dashboard_id": "…",
                            "display_type": "…",
                            "id": "…",
                            "layout_json": "…",
                            "query_sql": "…",
                            "query_version": "…",
                            "settings_json": "…",
                            "title": "…"
                          }
                        ],
                        "updated_at": 1.5,
                        "variables": [
                          {
                            "dashboard_id": "…",
                            "default_value": "…",
                            "kind": "…",
                            "name": "…",
                            "options_json": "…",
                            "source": "…"
                          }
                        ]
                      }
                    }
                  }
                },
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Dashboard"
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "summary": "import Dashboard Config",
        "tags": [
          "Dashboards"
        ]
      }
    },
    "/api/dashboards/reorder": {
      "post": {
        "operationId": "reorderDashboards",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "ids": {
                    "items": {
                      "type": "string"
                    },
                    "type": "array"
                  }
                },
                "required": [
                  "ids"
                ],
                "type": "object"
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "example": {
                    "summary": "Representative Spaniel response",
                    "value": {
                      "data": {
                        "ok": true
                      }
                    }
                  }
                },
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Ok"
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "summary": "reorder Dashboards",
        "tags": [
          "Dashboards"
        ]
      }
    },
...
    "/api/dashboards/{id}/config": {
      "get": {
        "operationId": "exportDashboardConfig",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/yaml": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "summary": "export Dashboard Config",
        "tags": [
          "Dashboards"
        ]
      }
    },
    "/api/dashboards/{id}/panels": {
      "post": {
        "operationId": "createDashboardPanel",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object"
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "example": {
                    "summary": "Representative Spaniel response",
                    "value": {
                      "data": {
                        "dashboard_id": "example",
                        "display_type": "single_value",
                        "id": "session_2026-10-07_1200",
                        "layout_json": "example",
                        "query_sql": "example",
                        "query_version": 1.5,
                        "settings_json": "example",
                        "title": "example"
                      }
                    }
                  }
                },
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/DashboardPanel"
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "summary": "create Dashboard Panel",
        "tags": [
          "Dashboards"
        ]
      }
    },
...
    "/api/dashboards/{id}/panels/{panelId}/move": {
      "post": {
        "operationId": "moveDashboardPanel",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "panelId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "direction": {
                    "enum": [
                      -1,
                      1
                    ],
                    "type": "integer"
                  }
                },
                "required": [
                  "direction"
                ],
                "type": "object"
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "example": {
                    "summary": "Representative Spaniel response",
                    "value": {
                      "data": {
                        "ok": true
                      }
                    }
                  }
                },
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Ok"
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "summary": "move Dashboard Panel",
        "tags": [
          "Dashboards"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Current Spaniel instance",
      "url": "/"
    }
  ],
...
}